+++ modules/septima-lang/tests/septima.spec.ts
@@ -729,6 +729,13 @@ describe('septima', () => {
     })
   })
   describe('undefined', () => {
+    // We want to verify that attributes with an undefined values do not exist in the object. To verify that we look
+    // at the keys of the object.
+    const keysOf = (u: unknown) => {
+      const casted = u as Record<string, unknown> // eslint-disable-line @typescript-eslint/consistent-type-assertions
+      return Object.keys(casted)
+    }
+
     test(`the 'undefined' literal evaluates to (a JS) undefined`, () => {
       expect(run(`let x = undefined; x`)).toBe(undefined)
     })
@@ -742,14 +749,15 @@ describe('septima', () => {
       expect(run(`['a', undefined, 'c']`)).toEqual(['a', undefined, 'c'])
     })
     test('an object attribute with a value of undefined is dropped from the object', () => {
-      // We want to verify that attributes with an undefined values do not exist in the object. To verify that we look
-      // at the keys of the object.
-      const keysOf = (u: unknown) => {
-        const casted = u as Record<string, unknown> // eslint-disable-line @typescript-eslint/consistent-type-assertions
-        return Object.keys(casted)
-      }
       expect(keysOf(run(`{n: 42, o: undefined, p: 'poo'}`))).toEqual(['n', 'p'])
       expect(keysOf(run(`Object.fromEntries([['n', 42], ['o', undefined], ['p', 'poo']])`))).toEqual(['n', 'p'])
+    })
+    test.todo('decide how overwriting with undefined works')
+    test('spreading an undefined in object is a no-op', () => {
+      expect(run(`{n: 42, ...undefined, p: 'poo'}`)).toEqual({ n: 42, p: 'poo' })
+    })
+    test('spreading an undefined in an array is a no-op', () => {
+      expect(run(`[42, ...undefined, 'poo']`)).toEqual([42, 'poo'])
     })
     test('produces a full trace when an undefined-reference-error is fired', () => {
       let message
@@ -794,8 +802,39 @@ describe('septima', () => {
         expect(run(`'Han' ?? 'Luke'`)).toEqual('Han')
       })
     })
-    test.todo('spread operator with undefined')
-    test.todo('decide how overwriting with undefined works')
+  })
+  describe('Casting functions', () => {
+    test('String()', () => {
+      expect(run(`String(42)`)).toEqual('42')
+      expect(run(`String("abc")`)).toEqual('abc')
+      expect(run(`String(true)`)).toEqual('true')
+      expect(run(`String(false)`)).toEqual('false')
+      expect(run(`String(undefined)`)).toEqual('undefined')
+      expect(run(`String({a: "alpha", b: [3,1,4], n: 42})`)).toEqual('{"a":"alpha","b":[3,1,4],"n":42}')
+      expect(run(`String(["abc", 3.14159, false, true, undefined])`)).toEqual('["abc",3.14159,false,true,null]')
+    })
+    test('Boolean()', () => {
+      expect(run(`Boolean(42)`)).toEqual(true)
+      expect(run(`Boolean(0)`)).toEqual(false)
+      expect(run(`Boolean("abc")`)).toEqual(true)
+      expect(run(`Boolean("")`)).toEqual(false)
+      expect(run(`Boolean(true)`)).toEqual(true)
+      expect(run(`Boolean(false)`)).toEqual(false)
+      expect(run(`Boolean(undefined)`)).toEqual(false)
+      expect(run(`Boolean({})`)).toEqual(true)
+      expect(run(`Boolean([])`)).toEqual(true)
+    })
+    test('Number()', () => {
+      expect(run(`Number(42)`)).toEqual(42)
+      expect(run(`Number(0)`)).toEqual(0)
+      expect(run(`Number("42")`)).toEqual(42)
+      expect(run(`Number("abc")`)).toEqual(NaN)
+      expect(run(`Number(true)`)).toEqual(1)
+      expect(run(`Number(false)`)).toEqual(0)
+      expect(run(`Number(undefined)`)).toEqual(NaN)
+      expect(run(`Number({})`)).toEqual(NaN)
+      expect(run(`Number([])`)).toEqual(NaN)
+    })
   })
   test.todo('support file names in locations')
   test.todo('string interpolation via `foo` strings')