+++ modules/septima-lang/change-log.md
@@ -9,6 +9,12 @@
 
 ### PR/164
 
-- undefined is now a first class value: `let x = undefined`
+- `undefined` is now a first class value: `let x = undefined`
 - `{a: undefined, n: 42}` is identical to `{n: 42}`
 - `undefined ?? 42` is `42`
+
+### PR/165
+
+- spreading `undefined` in an array is a no-op: `[42, ...undefined, 'poo']` is `[42, 'poo']`
+- spreading `undefined` in an object is a no-op: `{n: 42, ...undefined, p: 'poo'}` is `{n: 42, p: 'poo'}`
+- support `String(x)`, `Boolean(x)`, `Number(x)`